Added skybox and component support
Finally added suport for rendering skyboxes, for now the engine can only load cubemaps from an array of six images, but I plan of allowing LatLong maps.

Also finally moved to a scene system, to do this I fixed a lot of bugs with scene and scene queing and loading, now it posible to load Entire worlds without needing to create a layer.

Added the beginings of some important components, like the camera, directional and point light, and the Transform component.