Changed main rotoation system to euler
Changed the Transform component rotation system, previously, it used quaternions to store and manipulate rotation, but it has become too complicated to mantain. Now the Transform component uses Euler angles, and stores them as a 3-component vector.

In the future, I plan on adding quaternions back, but for now I don´t think it is worth the complication